bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.2
braceexpand==0.1.7
click==8.3.0
discord-ios==1.1.1
discord.py==2.6.3
frozenlist==1.7.0
idna==3.10
import_expression==2.2.1.post1
jishaku==2.6.0
loguru==0.7.3
multidict==6.6.4
pathlib==1.0.1
propcache==0.3.2
psutil==7.1.0
python-dotenv==1.1.1
redis==6.4.0
tabulate==0.9.0
typing_extensions==4.15.0
yarl==1.20.1
